role superadmin shouldnt allow destroy by API

+++ app/services/role/destroy_service.rb
@@ -3,6 +3,11 @@ class Role::DestroyService < ApplicationService
     role = Role.find(params[:id])
     raise RecordNotFound.new(params[:id], Role.model_name.human) if role.nil?
 
+    if Role.superadmin?(role)
+      role.errors.add(:base, 'role superadmin tidak boleh dihapus')
+      render_error_record(role)
+      return
+    end
     if role.destroy
       render_json({ message: "#{role.name} sukses dihapus" })
     else
